fix(client ts): TML-1442 fix missing json null types in browser.ys
