+++ packages/client-generator-ts/src/TSClient/file-generators/PrismaNamespaceBrowserFile.ts
@@ -2,6 +2,7 @@ import * as ts from '@prisma/ts-builders'
 
 import { Enum } from '../Enum'
 import { GenerateContext } from '../GenerateContext'
+import { nullTypes } from '../NullTypes'
 
 const jsDocHeader = `/*
  * WARNING: This is an internal file that is subject to change!
@@ -20,9 +21,14 @@ export function createPrismaNamespaceBrowserFile(context: GenerateContext): stri
 
   return `${jsDocHeader}
 ${ts.stringify(ts.moduleImport(`${context.runtimeBase}/index-browser`).asNamespace('runtime'))}
+
 export type * from '${context.importFileName(`../models`)}'
 export type * from '${context.importFileName(`./prismaNamespace`)}'
+
 export const Decimal = runtime.Decimal
+
+${nullTypes}
+
 ${new Enum(
   {
     name: 'ModelName',
@@ -30,9 +36,10 @@ ${new Enum(
   },
   true,
 ).toTS()}
-/**
+/*
  * Enums
  */
+
 ${prismaEnums?.join('\n\n')}
 `
 }
